Add tasks for McGonagall

+++ src/tasks.py
@@ -1,11 +1,15 @@
-from src.variables import local_deploy, channel_ids
+from src.functions import send_webhook, get_image
+from src.variables import local_deploy, channel_ids, system_embed_color
 
+from discord.embeds import Embed
+from discord.enums import EntityType, PrivacyLevel
 from discord.ext import tasks
 
-from datetime import datetime, time, timezone
+from datetime import datetime, time, timedelta, timezone
+import time as time_module
 import pytz
 
-__all__ = ["club_event_reminder", "midnight_reminder"] 
+__all__ = ["club_event_reminder", "game_midnight_reminder", "midnight_reminder", "game_reset_reminder"] 
 
 
 # SETTINGS 
@@ -19,7 +23,7 @@ time_trigger = {"game":       {"hour": 4,  "minute": 0,  "timezone": [pytz.timez
                 "midnight":   {"hour": 0,  "minute": 0,  "timezone": [pytz.timezone("Africa/Cairo"), timezone.utc]}, # UTC+2 / UTC
                 "club event": {"hour": 19, "minute": 25, "timezone": [timezone.utc]},}                               # UTC
 
-delete_after = time(hour=1, minute=0, second=0)
+delete_after = {"hours":1, "minutes":5, "seconds":0}
 
 weekday = {0:"Monday", 1:"Thursday",  2:"Wednesday", 3:"Tuesday", 4:"Friday", 5:"Saturday", 6:"Sunday"}
 
@@ -36,7 +40,7 @@ if test_events:
     time_trigger = {key:{"hour": hour, "minute": minute, "timezone": [tz]} for key in time_trigger}
     channel_ids = {key:channel_ids["testing"] for key in channel_ids}
     
-    delete_after = time(hour=0, minute=after_minutes, second=0)
+    delete_after = {"hours":0, "minutes":after_minutes*2, "seconds":0}
     
     del[hour, minute, tz]
 
@@ -47,15 +51,55 @@ idx = len(time_trigger["midnight"]["timezone"]) - 1
 @tasks.loop(time=time(hour=time_trigger["club event"]["hour"], 
                       minute=time_trigger["club event"]["minute"], 
                       tzinfo=time_trigger["club event"]["timezone"][0]))
-async def club_event_reminder():
+async def club_event_reminder(server):
     print(f'''"Club event" task running... {today}!''')
+
+    if test_events:
+        end_after = timedelta(minutes=after_minutes*2)
+        date = today + timedelta(minutes=after_minutes)
+    else:
+        end_after = timedelta(hours=1)
+        date = today.replace(hour=20, minute=30, second=0)
+    
+    unix_time_timer = convert_to_unix_time(date=date, mode="R")
+    unix_time_hour = convert_to_unix_time(date=date, mode="t")
+
+    event_info = {"title": "GOP Club Events!",
+                  "description": f"**We start {unix_time_timer}!**\nWe will begin with a Quiz, and after roughly 20 min we go over to a Dance Event!",
+                  "location": "HP: Magic Awakened  (Sphinx)"}
+    
+    url = "https://media.discordapp.net/attachments/1255614086033575977/1315444388515807292/template.png?ex=67576e8d&is=67561d0d&hm=ffffd3224bb60b5c3681c2017a43dc51238e3e30bafed9f4ad091977895be2d2"
+
+
+    # create event
+    await server.create_scheduled_event(name=event_info["title"],
+                                        start_time=date.astimezone(),
+                                        end_time=(date + end_after).astimezone(),
+                                        description=event_info["description"],
+                                        location=event_info["location"],
+                                        privacy_level=PrivacyLevel.guild_only,
+                                        entity_type=EntityType.external,
+                                        image=get_image(url=url))
+    
+    
+    # create notification message
+    embed = Embed(color=system_embed_color, title=event_info["title"], description=event_info["description"])
+    embed.set_author(icon_url="https://storage.googleapis.com/chronicle-assets/images/icons/bell-alert-white.png", name=f"Reminder: {weekday[date.weekday()]} Club Events!")
+    embed.add_field(name="Location", value=event_info["location"], inline=False)
+    embed.add_field(name="Scheduled for", value=f"{unix_time_hour}", inline=True)
+    embed.add_field(name="Duration", value="~1 hour", inline=True)
+
+
+    channel = server.get_channel(channel_ids["announcements"])
+    message = await send_webhook(target_channel=channel, user_name="Prof. McGonagall", content="<@&1278844289694171260>", embed=embed)
+    delete_message.start(message)
 
 
 # game midnight reminder:
 @tasks.loop(time=time(hour=time_trigger["midnight"]["hour"],
                       minute=time_trigger["midnight"]["minute"],
                       tzinfo=time_trigger["midnight"]["timezone"][0]))
-async def game_midnight_reminder():
+async def game_midnight_reminder(server):
     print(f'''"Game Midnight" task running... {today}!''')
 
 
@@ -78,11 +122,21 @@ async def midnight_reminder(server):
 @tasks.loop(time=time(hour=time_trigger["game"]["hour"],
                       minute=time_trigger["game"]["minute"],
                       tzinfo=time_trigger["game"]["timezone"][0]))
-async def game_reset_reminder():
+async def game_reset_reminder(server):
     print(f'''"Game Reset" task running... {today}!''')
 
 
 # delete message 
-@tasks.loop(time=delete_after) #Create the task
+@tasks.loop(hours=delete_after["hours"], minutes=delete_after["minutes"], seconds=delete_after["seconds"], count=2)
 async def delete_message(message):
-    await message.delete()
+    if delete_message.current_loop != 0:
+        print("Message deleted!")
+        await message.delete()
+
+
+def convert_to_unix_time(date: datetime, mode: str) -> str:
+    # get a tuple of the date attributes
+    date_tuple = (date.year, date.month, date.day, date.hour, date.minute, date.second)
+
+    # convert to unix time
+    return f'<t:{int(time_module.mktime(datetime(*date_tuple).timetuple()))}:{mode}>'
